requirements.txt
python-telegram-bot>=22.0,<23.0
Pillow>=10.0.0,<12.0.0
requests>=2.32.0,<3.0.0
